+++ cmd/neofs-lens/internal/meta/root.go
@@ -31,7 +31,6 @@ var Root = &cobra.Command{
 func init() {
 	Root.AddCommand(
 		listCMD,
-		listGraveyardCMD,
 		listGarbageCMD,
 		writeObjectCMD,
 		getCMD,