+++ pkg/local_object_storage/metabase/control.go
@@ -88,7 +88,6 @@ func (db *DB) init(reset bool) error {
 
 	mStaticBuckets := map[string]struct{}{
 		string(containerVolumeBucketName): {},
-		string(graveyardBucketName):       {},
 		string(toMoveItBucketName):        {},
 		string(garbageObjectsBucketName):  {},
 		string(shardInfoBucket):           {},